// src/api/generated/model/attemptResponseDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * TOEIC Dictation API
 * The TOEIC Dictation MVP API description
 * OpenAPI spec version: 1.0
 */
import type { AttemptDetailResponseDto } from './attemptDetailResponseDto';
import type { AttemptResponseDtoDifficultyMode } from './attemptResponseDtoDifficultyMode';

export interface AttemptResponseDto {
  id: string;
  user_id: string;
  lesson_id: string;
  difficulty_mode: AttemptResponseDtoDifficultyMode;
  score: number;
  accuracy_percent: number;
  correct_sentences: number;
  total_sentences: number;
  duration_seconds: number;
  started_at: string;
  completed_at: string;
  details?: AttemptDetailResponseDto[];
}